Moves mock module into snat_sim library
